refactor(rpc): extract ledger adapter
Refs #1483
