+++ internal/rpc/account_channels_test.go
@@ -70,7 +70,7 @@ func (m *mockAccountChannelsLedgerService) GetLedgerBySequence(seq uint32) (type
 func (m *mockAccountChannelsLedgerService) GetLedgerByHash(hash [32]byte) (types.LedgerReader, error) {
 	return accountQueryLedgerByHash(hash, m.validatedLedgerIndex)
 }
-func (m *mockAccountChannelsLedgerService) SubmitTransaction(txJSON []byte, txBlobHex ...string) (*types.SubmitResult, error) {
+func (m *mockAccountChannelsLedgerService) SubmitTransaction(txJSON []byte, txBlobHex string) (*types.SubmitResult, error) {
 	return nil, errors.New("not implemented")
 }
 func (m *mockAccountChannelsLedgerService) GetCurrentFees() (baseFee, reserveBase, reserveIncrement uint64) {
