--- internal/rpc/ledger_adapter.go
@@ -1,1184 +1,0 @@
-package rpc
-
-import (
-	"context"
-	"encoding/hex"
-	"encoding/json"
-	"fmt"
-	"strconv"
-	"time"
-
-	"github.com/LeJamon/go-xrpl/amendment"
-	binarycodec "github.com/LeJamon/go-xrpl/codec/binarycodec"
-	"github.com/LeJamon/go-xrpl/internal/ledger"
-	"github.com/LeJamon/go-xrpl/internal/ledger/service"
-	"github.com/LeJamon/go-xrpl/internal/ledger/state"
-	"github.com/LeJamon/go-xrpl/internal/rpc/handlers"
-	"github.com/LeJamon/go-xrpl/internal/rpc/types"
-	"github.com/LeJamon/go-xrpl/internal/tx"
-	"github.com/LeJamon/go-xrpl/internal/tx/mptutil"
-	"github.com/LeJamon/go-xrpl/internal/tx/ter"
-	"github.com/LeJamon/go-xrpl/protocol"
-	"github.com/LeJamon/go-xrpl/storage/relationaldb"
-)
-
-// LedgerServiceAdapter adapts the ledger service to the RPC LedgerService interface
-type LedgerServiceAdapter struct {
-	svc           *service.Service
-	txBroadcaster func(txBlob []byte) // called after successful submit to relay tx to peers
-}
-
-var _ types.LedgerService = (*LedgerServiceAdapter)(nil)
-var _ types.OwnerDirectoryReader = (*LedgerServiceAdapter)(nil)
-var _ types.TxTablesProvider = (*LedgerServiceAdapter)(nil)
-var _ types.RangedTransactionLookup = (*LedgerServiceAdapter)(nil)
-var _ types.TransactionRulesSource = (*LedgerServiceAdapter)(nil)
-
-// NewLedgerServiceAdapter creates a new adapter
-func NewLedgerServiceAdapter(svc *service.Service) *LedgerServiceAdapter {
-	return &LedgerServiceAdapter{svc: svc}
-}
-
-// SetTxBroadcaster sets the callback for relaying submitted transactions to P2P peers.
-// Called during server startup once the overlay is available.
-func (a *LedgerServiceAdapter) SetTxBroadcaster(fn func(txBlob []byte)) {
-	a.txBroadcaster = fn
-}
-
-// GetCurrentLedgerIndex returns the current open ledger index
-func (a *LedgerServiceAdapter) GetCurrentLedgerIndex() uint32 {
-	return a.svc.GetCurrentLedgerIndex()
-}
-
-// GetClosedLedgerIndex returns the last closed ledger index
-func (a *LedgerServiceAdapter) GetClosedLedgerIndex() uint32 {
-	return a.svc.GetClosedLedgerIndex()
-}
-
-// GetValidatedLedgerIndex returns the highest validated ledger index
-func (a *LedgerServiceAdapter) GetValidatedLedgerIndex() uint32 {
-	return a.svc.GetValidatedLedgerIndex()
-}
-
-// AcceptLedger closes the current open ledger (standalone mode only)
-func (a *LedgerServiceAdapter) AcceptLedger(ctx context.Context) (uint32, error) {
-	return a.svc.AcceptLedger(ctx)
-}
-
-// AcceptLedgerAt is AcceptLedger with an explicit close_time.
-func (a *LedgerServiceAdapter) AcceptLedgerAt(ctx context.Context, closeTime time.Time) (uint32, error) {
-	return a.svc.AcceptLedgerAt(ctx, closeTime)
-}
-
-// IsStandalone returns true if running in standalone mode
-func (a *LedgerServiceAdapter) IsStandalone() bool {
-	return a.svc.IsStandalone()
-}
-
-// GetServerInfo returns server status information
-func (a *LedgerServiceAdapter) GetServerInfo() types.LedgerServerInfo {
-	info := a.svc.GetServerInfo()
-	return types.LedgerServerInfo{
-		Standalone:               info.Standalone,
-		ServerState:              info.ServerState,
-		NeedsNetworkLedger:       info.NeedsNetworkLedger,
-		OpenLedgerSeq:            info.OpenLedgerSeq,
-		ClosedLedgerSeq:          info.ClosedLedgerSeq,
-		ClosedLedgerHash:         info.ClosedLedgerHash,
-		ClosedLedgerCloseTime:    info.ClosedLedgerCloseTime,
-		HaveValidated:            info.HaveValidated,
-		ValidatedLedgerSeq:       info.ValidatedLedgerSeq,
-		ValidatedLedgerHash:      info.ValidatedLedgerHash,
-		ValidatedLedgerCloseTime: info.ValidatedLedgerCloseTime,
-		CompleteLedgers:          info.CompleteLedgers,
-		HavePublished:            info.HavePublished,
-		PublishedLedgerSeq:       info.PublishedLedgerSeq,
-		NetworkID:                info.NetworkID,
-	}
-}
-
-// GetLedgerBySequence returns a ledger by its sequence number
-func (a *LedgerServiceAdapter) GetLedgerBySequence(seq uint32) (types.LedgerReader, error) {
-	l, err := a.svc.GetLedgerBySequence(seq)
-	if err != nil {
-		return nil, err
-	}
-	if l.IsOpen() {
-		l, err = l.Snapshot()
-		if err != nil {
-			return nil, err
-		}
-	}
-	return &ledgerReaderAdapter{l: l}, nil
-}
-
-// GetLedgerByHash returns a ledger by its hash
-func (a *LedgerServiceAdapter) GetLedgerByHash(hash [32]byte) (types.LedgerReader, error) {
-	l, err := a.svc.GetLedgerByHash(hash)
-	if err != nil {
-		return nil, err
-	}
-	return &ledgerReaderAdapter{l: l}, nil
-}
-
-func (a *LedgerServiceAdapter) GetLedgerByHashContext(ctx context.Context, hash [32]byte) (types.LedgerReader, error) {
-	l, err := a.svc.GetLedgerByHashContext(ctx, hash)
-	if err != nil {
-		return nil, err
-	}
-	return &ledgerReaderAdapter{l: l}, nil
-}
-
-// GetGenesisAccount returns the genesis account address
-func (a *LedgerServiceAdapter) GetGenesisAccount() (string, error) {
-	return a.svc.GetGenesisAccount()
-}
-
-// ledgerReaderAdapter adapts ledger.Ledger to types.LedgerReader interface
-type ledgerReaderAdapter struct {
-	l *ledger.Ledger
-}
-
-func (a *ledgerReaderAdapter) Sequence() uint32 {
-	return a.l.Sequence()
-}
-
-func (a *ledgerReaderAdapter) Hash() [32]byte {
-	return a.l.Hash()
-}
-
-func (a *ledgerReaderAdapter) ParentHash() [32]byte {
-	return a.l.ParentHash()
-}
-
-func (a *ledgerReaderAdapter) IsClosed() bool {
-	return a.l.IsClosed()
-}
-
-func (a *ledgerReaderAdapter) IsValidated() bool {
-	return a.l.IsValidated()
-}
-
-func (a *ledgerReaderAdapter) TotalDrops() uint64 {
-	return a.l.TotalDrops()
-}
-
-func (a *ledgerReaderAdapter) CloseTime() int64 {
-	return protocol.RippleSeconds(a.l.CloseTime())
-}
-
-func (a *ledgerReaderAdapter) CloseTimeResolution() uint32 {
-	return uint32(a.l.Header().CloseTimeResolution)
-}
-
-func (a *ledgerReaderAdapter) CloseFlags() uint8 {
-	return a.l.Header().CloseFlags
-}
-
-func (a *ledgerReaderAdapter) ParentCloseTime() int64 {
-	return protocol.RippleSeconds(a.l.ParentCloseTime())
-}
-
-func (a *ledgerReaderAdapter) TxMapHash() [32]byte {
-	return a.l.Header().TxHash
-}
-
-func (a *ledgerReaderAdapter) StateMapHash() [32]byte {
-	return a.l.Header().AccountHash
-}
-
-func (a *ledgerReaderAdapter) ForEachTransaction(fn func(txHash [32]byte, txData []byte) bool) error {
-	return a.l.ForEachTransaction(fn)
-}
-
-func (a *ledgerReaderAdapter) ForEachTransactionContext(ctx context.Context, fn func(txHash [32]byte, txData []byte) bool) error {
-	return a.l.ForEachTransactionContext(ctx, fn)
-}
-
-func (a *ledgerReaderAdapter) GetLedgerTransaction(txHash [32]byte) ([]byte, bool, error) {
-	return a.l.GetTransaction(txHash)
-}
-
-func (a *ledgerReaderAdapter) GetLedgerTransactionContext(ctx context.Context, txHash [32]byte) ([]byte, bool, error) {
-	return a.l.GetTransactionContext(ctx, txHash)
-}
-
-func (a *ledgerReaderAdapter) ForEachLedgerStateContext(ctx context.Context, fn func(key [32]byte, data []byte) bool) error {
-	return a.l.ForEachContext(ctx, fn)
-}
-
-func (a *ledgerReaderAdapter) LedgerAmendmentRules() *amendment.Rules {
-	return a.l.Rules()
-}
-
-func (a *ledgerReaderAdapter) LedgerAmendmentRulesWithError() (*amendment.Rules, error) {
-	rules := a.l.Rules()
-	if rules == nil {
-		return nil, fmt.Errorf("ledger %d amendment rules unavailable", a.l.Sequence())
-	}
-	return rules, nil
-}
-
-// SubmitTransaction submits a transaction to the open ledger.
-// The optional txBlobHex is the original signed transaction blob in hex.
-// This is used for canonical re-ordering during AcceptLedger to ensure
-// the exact same bytes (and thus same tx hash) are used during re-application.
-func (a *LedgerServiceAdapter) SubmitTransaction(txJSON []byte, txBlobHex ...string) (*types.SubmitResult, error) {
-	var blobHex string
-	if len(txBlobHex) > 0 {
-		blobHex = txBlobHex[0]
-	}
-	return a.submitTransaction(txJSON, blobHex, false)
-}
-
-// SubmitTransactionFailHard mirrors SubmitTransaction with rippled's
-// tapFAIL_HARD semantics: a tx that does not apply is not held in
-// localTxs, not relayed, and not pushed onto the canonical pendingTxs
-// pool. Submit handlers route here when the client sent fail_hard:true.
-func (a *LedgerServiceAdapter) SubmitTransactionFailHard(txJSON []byte, txBlobHex string) (*types.SubmitResult, error) {
-	return a.submitTransaction(txJSON, txBlobHex, true)
-}
-
-// errorSubmitResult builds a non-applied SubmitResult whose engine fields are
-// sourced from the canonical TER table, so the result token, code, and message
-// stay in lockstep with rippled.
-func errorSubmitResult(r ter.Result) *types.SubmitResult {
-	return &types.SubmitResult{
-		EngineResult:        r.String(),
-		EngineResultCode:    int(r),
-		EngineResultMessage: r.Message(),
-		Applied:             false,
-	}
-}
-
-func malformedSubmitResult() *types.SubmitResult {
-	return errorSubmitResult(ter.TemMALFORMED)
-}
-
-func internalSubmitResult() *types.SubmitResult {
-	return errorSubmitResult(ter.TefINTERNAL)
-}
-
-func (a *LedgerServiceAdapter) submitTransaction(txJSON []byte, txBlobHex string, failHard bool) (*types.SubmitResult, error) {
-	// Parse the transaction from JSON
-	transaction, err := tx.ParseJSON(txJSON)
-	if err != nil {
-		return malformedSubmitResult(), nil
-	}
-
-	// Use the original signed blob if provided, otherwise re-encode
-	var rawBlob []byte
-	if txBlobHex != "" {
-		var decodeErr error
-		rawBlob, decodeErr = hex.DecodeString(txBlobHex)
-		if decodeErr != nil {
-			return nil, fmt.Errorf("decode tx_blob: %w", decodeErr)
-		}
-	}
-	if rawBlob == nil {
-		if txMap, fErr := transaction.Flatten(); fErr == nil {
-			if hexStr, eErr := binarycodec.Encode(txMap); eErr == nil {
-				var decodeErr error
-				rawBlob, decodeErr = hex.DecodeString(hexStr)
-				if decodeErr != nil {
-					return nil, fmt.Errorf("decode encoded transaction: %w", decodeErr)
-				}
-			}
-		}
-	}
-	if rawBlob == nil {
-		var jsonMap map[string]any
-		if jErr := json.Unmarshal(txJSON, &jsonMap); jErr == nil {
-			if hexStr, eErr := binarycodec.Encode(jsonMap); eErr == nil {
-				var decodeErr error
-				rawBlob, decodeErr = hex.DecodeString(hexStr)
-				if decodeErr != nil {
-					return nil, fmt.Errorf("decode encoded transaction: %w", decodeErr)
-				}
-			}
-		}
-	}
-
-	// Submit to the service with the raw blob for canonical ordering
-	result, err := a.svc.SubmitTransaction(transaction, rawBlob, failHard)
-	if err != nil {
-		return internalSubmitResult(), nil
-	}
-
-	// Relay only what rippled relays. Mirrors NetworkOPs.cpp:1685-1689:
-	//
-	//     (e.applied || ((mMode != FULL) && !failType && e.local) ||
-	//      e.result == terQUEUED) && !enforceFailHard
-	//
-	// where `e.applied = isTesSuccess(ret.ter) || isTecClaim(ret.ter)`
-	// (apply.cpp:196,206). On-the-wire set: tesSUCCESS, tec*, terQUEUED.
-	// All other ter* codes are held locally (addHeldTransaction) and
-	// must NOT be relayed — peers would just re-fail with the same
-	// retry code. fail_hard suppresses relay on non-apply (matches
-	// NetworkOPs.cpp:1688 `&& !enforceFailHard`).
-	broadcast := false
-	relayable := result.Applied || result.Result == ter.TerQUEUED
-	if failHard && !result.Applied {
-		relayable = false
-	}
-	if relayable && rawBlob != nil && a.txBroadcaster != nil {
-		a.txBroadcaster(rawBlob)
-		broadcast = true
-	}
-
-	// Kept mirrors rippled's setKept (NetworkOPs.cpp:1674-1683): a local
-	// submission is held in the LocalTxs pool — and thus "kept" — whenever
-	// addLocal && !enforceFailHard, i.e. (!fail_hard || result==tesSUCCESS).
-	// rippled does not filter by TER on the local-push path, so tef/tem/tel
-	// are kept too (they age out of LocalTxs after at most 5 ledgers). This
-	// is the exact condition Service.SubmitTransaction uses for the localTxs
-	// push, so the wire value reflects the actual held-pool decision.
-	queued := result.Result == ter.TerQUEUED
-	kept := (!failHard || result.Result == ter.TesSUCCESS) &&
-		result.Result != ter.TefALREADY
-
-	return &types.SubmitResult{
-		EngineResult:           result.Result.String(),
-		EngineResultCode:       int(result.Result),
-		EngineResultMessage:    result.Message,
-		Applied:                result.Applied,
-		Broadcast:              broadcast,
-		Queued:                 queued,
-		Kept:                   kept,
-		Fee:                    result.Fee,
-		CurrentLedger:          result.CurrentLedger,
-		CurrentLedgerCloseTime: result.CurrentLedgerCloseTime,
-		ValidatedLedger:        result.ValidatedLedger,
-	}, nil
-}
-
-// GetCurrentFees returns the current fee settings
-func (a *LedgerServiceAdapter) GetCurrentFees() (baseFee, reserveBase, reserveIncrement uint64) {
-	return a.svc.GetCurrentFees()
-}
-
-// GetAccountInfo retrieves account information from the ledger
-func (a *LedgerServiceAdapter) GetAccountInfo(ctx context.Context, account string, ledgerIndex string) (*types.AccountInfo, error) {
-	result, err := a.svc.GetAccountInfo(ctx, account, ledgerIndex)
-	if err != nil {
-		return nil, err
-	}
-
-	var prevTxnID string
-	zeroHash := [32]byte{}
-	if result.PreviousTxnID != zeroHash {
-		prevTxnID = handlers.FormatLedgerHash(result.PreviousTxnID)
-	}
-
-	return &types.AccountInfo{
-		Account:           result.Account,
-		Balance:           strconv.FormatUint(result.Balance, 10),
-		Flags:             result.Flags,
-		OwnerCount:        result.OwnerCount,
-		Sequence:          result.Sequence,
-		RegularKey:        result.RegularKey,
-		Domain:            result.Domain,
-		EmailHash:         result.EmailHash,
-		TransferRate:      result.TransferRate,
-		TickSize:          result.TickSize,
-		PreviousTxnID:     prevTxnID,
-		PreviousTxnLgrSeq: result.PreviousTxnLgrSeq,
-		LedgerIndex:       result.LedgerIndex,
-		LedgerHash:        handlers.FormatLedgerHash(result.LedgerHash),
-		Validated:         result.Validated,
-		RawData:           result.RawData,
-		Index:             handlers.FormatLedgerHash(result.Index),
-	}, nil
-}
-
-// GetTransaction retrieves a transaction by its hash
-func (a *LedgerServiceAdapter) GetTransaction(txHash [32]byte) (*types.TransactionInfo, error) {
-	result, err := a.svc.GetTransaction(txHash)
-	if err != nil {
-		return nil, err
-	}
-	return rpcTransactionInfo(result), nil
-}
-
-// GetTransactionWithRange performs the optional transaction-table lookup used
-// by the tx RPC without adding that method to the broad ledger service contract.
-func (a *LedgerServiceAdapter) GetTransactionWithRange(ctx context.Context, txHash [32]byte, minLedger, maxLedger uint32) (*types.TransactionInfo, types.TxSearchResult, error) {
-	result, searched, err := a.svc.GetTransactionWithRange(ctx, txHash, minLedger, maxLedger)
-	if result == nil {
-		return nil, rpcTxSearchResult(searched), err
-	}
-	return rpcTransactionInfo(result), rpcTxSearchResult(searched), err
-}
-
-func rpcTransactionInfo(result *service.TransactionResult) *types.TransactionInfo {
-	ledgerHash := ""
-	if result.LedgerHash != ([32]byte{}) {
-		ledgerHash = handlers.FormatLedgerHash(result.LedgerHash)
-	}
-	return &types.TransactionInfo{
-		TxData:      result.TxData,
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  ledgerHash,
-		Validated:   result.Validated,
-		TxIndex:     result.TxIndex,
-		CloseTime:   result.CloseTime,
-	}
-}
-
-func (a *LedgerServiceAdapter) SearchTransaction(ctx context.Context, txHash [32]byte, ledgerRange *types.TransactionSearchRange) (*types.TransactionSearchResult, error) {
-	var serviceRange *relationaldb.LedgerRange
-	if ledgerRange != nil {
-		serviceRange = &relationaldb.LedgerRange{
-			Min: relationaldb.LedgerIndex(ledgerRange.Min),
-			Max: relationaldb.LedgerIndex(ledgerRange.Max),
-		}
-	}
-	result, err := a.svc.SearchTransaction(ctx, txHash, serviceRange)
-	if err != nil {
-		return nil, err
-	}
-	response := &types.TransactionSearchResult{}
-	if result.Transaction != nil {
-		response.Transaction = rpcTransactionInfo(result.Transaction)
-	}
-	switch result.Searched {
-	case relationaldb.TxSearchAll:
-		searchedAll := true
-		response.SearchedAll = &searchedAll
-	case relationaldb.TxSearchSome:
-		searchedAll := false
-		response.SearchedAll = &searchedAll
-	}
-	return response, nil
-}
-
-func (a *LedgerServiceAdapter) GetLedgerContext(ctx context.Context, sequence uint32) (*types.LedgerContext, error) {
-	result, err := a.svc.GetLedgerContext(ctx, sequence)
-	if err != nil {
-		return nil, err
-	}
-	return &types.LedgerContext{Hash: result.Hash, CloseTime: result.CloseTime}, nil
-}
-
-func rpcTxSearchResult(result relationaldb.TxSearchResult) types.TxSearchResult {
-	switch result {
-	case relationaldb.TxSearchSome:
-		return types.TxSearchSome
-	case relationaldb.TxSearchAll:
-		return types.TxSearchAll
-	default:
-		return types.TxSearchUnknown
-	}
-}
-
-// StoreTransaction stores a transaction in the current ledger
-func (a *LedgerServiceAdapter) StoreTransaction(txHash [32]byte, txData []byte) error {
-	return a.svc.StoreTransaction(txHash, txData)
-}
-
-// GetAccountLines retrieves trust lines for an account
-func (a *LedgerServiceAdapter) GetAccountLines(ctx context.Context, account string, ledgerIndex string, peer string, limit uint32, marker string) (*types.AccountLinesResult, error) {
-	result, err := a.svc.GetAccountLines(ctx, account, ledgerIndex, peer, limit, marker)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service types to RPC types
-	lines := make([]types.TrustLine, len(result.Lines))
-	for i, line := range result.Lines {
-		lines[i] = types.TrustLine{
-			Account:        line.Account,
-			Balance:        line.Balance,
-			Currency:       line.Currency,
-			Limit:          line.Limit,
-			LimitPeer:      line.LimitPeer,
-			QualityIn:      line.QualityIn,
-			QualityOut:     line.QualityOut,
-			NoRipple:       line.NoRipple,
-			NoRipplePeer:   line.NoRipplePeer,
-			Authorized:     line.Authorized,
-			PeerAuthorized: line.PeerAuthorized,
-			Freeze:         line.Freeze,
-			FreezePeer:     line.FreezePeer,
-			DeepFreeze:     line.DeepFreeze,
-			DeepFreezePeer: line.DeepFreezePeer,
-			HasReserve:     line.HasReserve,
-		}
-	}
-
-	return &types.AccountLinesResult{
-		Account:     result.Account,
-		Lines:       lines,
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Validated:   result.Validated,
-		Marker:      result.Marker,
-	}, nil
-}
-
-// GetAccountOffers retrieves offers for an account
-func (a *LedgerServiceAdapter) GetAccountOffers(ctx context.Context, account string, ledgerIndex string, limit uint32, marker string) (*types.AccountOffersResult, error) {
-	result, err := a.svc.GetAccountOffers(ctx, account, ledgerIndex, limit, marker)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service types to RPC types
-	offers := make([]types.AccountOffer, len(result.Offers))
-	for i, offer := range result.Offers {
-		offers[i] = types.AccountOffer{
-			Flags:      offer.Flags,
-			Seq:        offer.Seq,
-			TakerGets:  offer.TakerGets,
-			TakerPays:  offer.TakerPays,
-			Quality:    offer.Quality,
-			Expiration: offer.Expiration,
-		}
-	}
-
-	return &types.AccountOffersResult{
-		Account:     result.Account,
-		Offers:      offers,
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Validated:   result.Validated,
-		Marker:      result.Marker,
-	}, nil
-}
-
-// GetBookOffers retrieves offers from an order book
-func (a *LedgerServiceAdapter) GetBookOffers(ctx context.Context, takerGets, takerPays types.Amount, taker, domain string, ledgerIndex string, limit uint32, marker string, withProofs bool) (*types.BookOffersResult, error) {
-	txTakerGets, err := rpcBookAmount(takerGets)
-	if err != nil {
-		return nil, err
-	}
-	txTakerPays, err := rpcBookAmount(takerPays)
-	if err != nil {
-		return nil, err
-	}
-
-	result, err := a.svc.GetBookOffers(ctx, txTakerGets, txTakerPays, taker, domain, ledgerIndex, limit, marker, withProofs)
-	if err != nil {
-		return nil, err
-	}
-
-	offers := make([]types.BookOffer, len(result.Offers))
-	for i, offer := range result.Offers {
-		offers[i] = types.BookOffer{
-			Account:           offer.Account,
-			BookDirectory:     offer.BookDirectory,
-			BookNode:          offer.BookNode,
-			Expiration:        offer.Expiration,
-			Flags:             offer.Flags,
-			LedgerEntryType:   offer.LedgerEntryType,
-			OwnerNode:         offer.OwnerNode,
-			PreviousTxnID:     offer.PreviousTxnID,
-			PreviousTxnLgrSeq: offer.PreviousTxnLgrSeq,
-			Sequence:          offer.Sequence,
-			TakerGets:         offer.TakerGets,
-			TakerPays:         offer.TakerPays,
-			DomainID:          offer.DomainID,
-			AdditionalBooks:   offer.AdditionalBooks,
-			Index:             offer.Index,
-			Quality:           offer.Quality,
-			OwnerFunds:        offer.OwnerFunds,
-			TakerGetsFunded:   offer.TakerGetsFunded,
-			TakerPaysFunded:   offer.TakerPaysFunded,
-			Proof:             offer.Proof,
-		}
-	}
-
-	return &types.BookOffersResult{
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Offers:      offers,
-		Validated:   result.Validated,
-		Marker:      result.Marker,
-	}, nil
-}
-
-func rpcBookAmount(amount types.Amount) (tx.Amount, error) {
-	if amount.IsMPT() {
-		id, err := mptutil.DecodeID(amount.MPTIssuanceID)
-		if err != nil {
-			return tx.Amount{}, fmt.Errorf("invalid mpt issuance id: %w", err)
-		}
-		issuer := state.EncodeAccountIDSafe(mptutil.Issuer(id))
-		return state.NewMPTAmountWithIssuanceID(0, issuer, mptutil.EncodeID(id)), nil
-	}
-	if amount.Currency == "" || amount.Currency == "XRP" {
-		return tx.NewXRPAmount(0), nil
-	}
-	return tx.NewIssuedAmountFromFloat64(0, amount.Currency, amount.Issuer), nil
-}
-
-// UseTxTables implements types.TxTablesProvider.
-func (a *LedgerServiceAdapter) UseTxTables() bool {
-	return a.svc.UseTxTables()
-}
-
-// GetAccountTransactions retrieves transaction history for an account
-func (a *LedgerServiceAdapter) GetAccountTransactions(ctx context.Context, account string, ledgerMin, ledgerMax int64, limit uint32, marker *types.AccountTxMarker, forward bool) (*types.AccountTxResult, error) {
-	// Convert RPC marker to service marker
-	var svcMarker *relationaldb.AccountTxMarker
-	if marker != nil {
-		svcMarker = &relationaldb.AccountTxMarker{
-			LedgerSeq: relationaldb.LedgerIndex(marker.LedgerSeq),
-			TxnSeq:    marker.TxnSeq,
-		}
-	}
-
-	result, err := a.svc.GetAccountTransactions(ctx, account, ledgerMin, ledgerMax, limit, svcMarker, forward)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	txs := make([]types.AccountTransaction, len(result.Transactions))
-	for i, tx := range result.Transactions {
-		txs[i] = types.AccountTransaction{
-			Hash:        tx.Hash,
-			LedgerIndex: tx.LedgerIndex,
-			TxnSeq:      tx.TxnSeq,
-			TxBlob:      tx.TxBlob,
-			Meta:        tx.Meta,
-		}
-	}
-
-	var rpcMarker *types.AccountTxMarker
-	if result.Marker != nil {
-		rpcMarker = &types.AccountTxMarker{
-			LedgerSeq: uint32(result.Marker.LedgerSeq),
-			TxnSeq:    result.Marker.TxnSeq,
-		}
-	}
-
-	return &types.AccountTxResult{
-		Account:      result.Account,
-		LedgerMin:    result.LedgerMin,
-		LedgerMax:    result.LedgerMax,
-		Limit:        result.Limit,
-		Marker:       rpcMarker,
-		Transactions: txs,
-		Validated:    result.Validated,
-	}, nil
-}
-
-// GetTransactionHistory retrieves recent transactions
-func (a *LedgerServiceAdapter) GetTransactionHistory(ctx context.Context, startIndex uint32) (*types.TxHistoryResult, error) {
-	result, err := a.svc.GetTransactionHistory(ctx, startIndex)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	txs := make([]types.AccountTransaction, len(result.Transactions))
-	for i, tx := range result.Transactions {
-		txs[i] = types.AccountTransaction{
-			Hash:        tx.Hash,
-			LedgerIndex: tx.LedgerIndex,
-			TxBlob:      tx.TxBlob,
-			Meta:        tx.Meta,
-		}
-	}
-
-	return &types.TxHistoryResult{
-		Index:        result.Index,
-		Transactions: txs,
-	}, nil
-}
-
-// GetLedgerRange retrieves ledger hashes for a range of sequences
-func (a *LedgerServiceAdapter) GetLedgerRange(ctx context.Context, minSeq, maxSeq uint32) (*types.LedgerRangeResult, error) {
-	result, err := a.svc.GetLedgerRange(ctx, minSeq, maxSeq)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.LedgerRangeResult{
-		LedgerFirst: result.LedgerFirst,
-		LedgerLast:  result.LedgerLast,
-		Hashes:      result.Hashes,
-	}, nil
-}
-
-// GetLedgerEntry retrieves a specific ledger entry by its index/key
-func (a *LedgerServiceAdapter) GetLedgerEntry(ctx context.Context, entryKey [32]byte, ledgerIndex string) (*types.LedgerEntryResult, error) {
-	result, err := a.svc.GetLedgerEntry(ctx, entryKey, ledgerIndex)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.LedgerEntryResult{
-		Index:       result.Index,
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Node:        result.Node,
-		NodeBinary:  handlers.FormatHash(result.Node),
-		Validated:   result.Validated,
-	}, nil
-}
-
-// GetLedgerData retrieves all ledger state entries with pagination
-func (a *LedgerServiceAdapter) GetLedgerData(ctx context.Context, ledgerIndex string, limit uint32, marker string) (*types.LedgerDataResult, error) {
-	result, err := a.svc.GetLedgerData(ctx, ledgerIndex, limit, marker)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	state := make([]types.LedgerDataItem, len(result.State))
-	for i, item := range result.State {
-		state[i] = types.LedgerDataItem{
-			Index: item.Index,
-			Data:  item.Data,
-		}
-	}
-
-	rpcResult := &types.LedgerDataResult{
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		State:       state,
-		Marker:      result.Marker,
-		Validated:   result.Validated,
-	}
-
-	// Convert ledger header info if present
-	if result.LedgerHeader != nil {
-		rpcResult.LedgerHeader = &types.LedgerHeaderInfo{
-			AccountHash:         result.LedgerHeader.AccountHash,
-			CloseFlags:          result.LedgerHeader.CloseFlags,
-			CloseTime:           result.LedgerHeader.CloseTime,
-			CloseTimeHuman:      result.LedgerHeader.CloseTimeHuman,
-			CloseTimeISO:        result.LedgerHeader.CloseTimeISO,
-			CloseTimeResolution: result.LedgerHeader.CloseTimeResolution,
-			Closed:              result.LedgerHeader.Closed,
-			LedgerHash:          result.LedgerHeader.LedgerHash,
-			LedgerIndex:         result.LedgerHeader.LedgerIndex,
-			ParentCloseTime:     result.LedgerHeader.ParentCloseTime,
-			ParentHash:          result.LedgerHeader.ParentHash,
-			TotalCoins:          result.LedgerHeader.TotalCoins,
-			TransactionHash:     result.LedgerHeader.TransactionHash,
-		}
-	}
-
-	return rpcResult, nil
-}
-
-// GetAccountObjects retrieves all objects owned by an account
-func (a *LedgerServiceAdapter) GetAccountObjects(ctx context.Context, account string, ledgerIndex string, objType string, limit uint32, marker string) (*types.AccountObjectsResult, error) {
-	result, err := a.svc.GetAccountObjects(ctx, account, ledgerIndex, objType, limit, marker)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	objects := make([]types.AccountObjectItem, len(result.AccountObjects))
-	for i, obj := range result.AccountObjects {
-		objects[i] = types.AccountObjectItem{
-			Index:           obj.Index,
-			LedgerEntryType: obj.LedgerEntryType,
-			Data:            obj.Data,
-		}
-	}
-
-	return &types.AccountObjectsResult{
-		Account:        result.Account,
-		AccountObjects: objects,
-		LedgerIndex:    result.LedgerIndex,
-		LedgerHash:     result.LedgerHash,
-		Validated:      result.Validated,
-		Marker:         result.Marker,
-	}, nil
-}
-
-// GetOwnerInfo walks the account's owner directory for the owner_info RPC,
-// implementing types.OwnerDirectoryReader.
-func (a *LedgerServiceAdapter) GetOwnerInfo(ctx context.Context, account string, ledgerIndex string) (*types.OwnerInfoResult, error) {
-	result, err := a.svc.GetOwnerInfo(ctx, account, ledgerIndex)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.OwnerInfoResult{
-		Offers:      toRPCAccountObjectItems(result.Offers),
-		RippleLines: toRPCAccountObjectItems(result.RippleLines),
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Validated:   result.Validated,
-	}, nil
-}
-
-func toRPCAccountObjectItems(items []service.AccountObjectItem) []types.AccountObjectItem {
-	out := make([]types.AccountObjectItem, len(items))
-	for i, obj := range items {
-		out[i] = types.AccountObjectItem{
-			Index:           obj.Index,
-			LedgerEntryType: obj.LedgerEntryType,
-			Data:            obj.Data,
-		}
-	}
-	return out
-}
-
-// GetAccountChannels retrieves payment channels for an account
-func (a *LedgerServiceAdapter) GetAccountChannels(ctx context.Context, account string, destinationAccount string, ledgerIndex string, limit uint32, marker string) (*types.AccountChannelsResult, error) {
-	result, err := a.svc.GetAccountChannels(ctx, account, destinationAccount, ledgerIndex, limit, marker)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	channels := make([]types.AccountChannel, len(result.Channels))
-	for i, ch := range result.Channels {
-		channels[i] = types.AccountChannel{
-			ChannelID:          ch.ChannelID,
-			Account:            ch.Account,
-			DestinationAccount: ch.DestinationAccount,
-			Amount:             ch.Amount,
-			Balance:            ch.Balance,
-			SettleDelay:        ch.SettleDelay,
-			PublicKey:          ch.PublicKey,
-			PublicKeyHex:       ch.PublicKeyHex,
-			Expiration:         ch.Expiration,
-			CancelAfter:        ch.CancelAfter,
-			SourceTag:          ch.SourceTag,
-			DestinationTag:     ch.DestinationTag,
-			HasSourceTag:       ch.HasSourceTag,
-			HasDestTag:         ch.HasDestTag,
-		}
-	}
-
-	return &types.AccountChannelsResult{
-		Account:     result.Account,
-		Channels:    channels,
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Validated:   result.Validated,
-		Marker:      result.Marker,
-	}, nil
-}
-
-// GetAccountCurrencies retrieves currencies an account can send and receive
-func (a *LedgerServiceAdapter) GetAccountCurrencies(ctx context.Context, account string, ledgerIndex string) (*types.AccountCurrenciesResult, error) {
-	result, err := a.svc.GetAccountCurrencies(ctx, account, ledgerIndex)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.AccountCurrenciesResult{
-		ReceiveCurrencies: result.ReceiveCurrencies,
-		SendCurrencies:    result.SendCurrencies,
-		LedgerIndex:       result.LedgerIndex,
-		LedgerHash:        result.LedgerHash,
-		Validated:         result.Validated,
-	}, nil
-}
-
-func (a *LedgerServiceAdapter) GetAccountNFTs(ctx context.Context, account string, ledgerIndex string, limit uint32, marker string) (*types.AccountNFTsResult, error) {
-	result, err := a.svc.GetAccountNFTs(ctx, account, ledgerIndex, limit, marker)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	nfts := make([]types.NFTInfo, len(result.AccountNFTs))
-	for i, nft := range result.AccountNFTs {
-		nfts[i] = types.NFTInfo{
-			Flags:        nft.Flags,
-			Issuer:       nft.Issuer,
-			NFTokenID:    nft.NFTokenID,
-			NFTokenTaxon: nft.NFTokenTaxon,
-			URI:          nft.URI,
-			NFTSerial:    nft.NFTSerial,
-			TransferFee:  nft.TransferFee,
-		}
-	}
-
-	return &types.AccountNFTsResult{
-		Account:     result.Account,
-		AccountNFTs: nfts,
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Validated:   result.Validated,
-		Marker:      result.Marker,
-	}, nil
-}
-
-// GetNoRippleCheck checks trust lines for proper NoRipple flag settings
-func (a *LedgerServiceAdapter) GetNoRippleCheck(ctx context.Context, account string, role string, ledgerIndex string, limit uint32, transactions bool) (*types.NoRippleCheckResult, error) {
-	result, err := a.svc.GetNoRippleCheck(ctx, account, role, ledgerIndex, limit, transactions)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	var txs []types.SuggestedTransaction
-	if len(result.Transactions) > 0 {
-		txs = make([]types.SuggestedTransaction, len(result.Transactions))
-		for i, tx := range result.Transactions {
-			txs[i] = types.SuggestedTransaction{
-				TransactionType: tx.TransactionType,
-				Account:         tx.Account,
-				Fee:             tx.Fee,
-				Sequence:        tx.Sequence,
-				SetFlag:         tx.SetFlag,
-				Flags:           tx.Flags,
-				LimitAmount:     tx.LimitAmount,
-			}
-		}
-	}
-
-	return &types.NoRippleCheckResult{
-		Problems:     result.Problems,
-		Transactions: txs,
-		LedgerIndex:  result.LedgerIndex,
-		LedgerHash:   result.LedgerHash,
-		Validated:    result.Validated,
-	}, nil
-}
-
-// GetGatewayBalances retrieves obligations and balances for a gateway account
-func (a *LedgerServiceAdapter) GetGatewayBalances(ctx context.Context, account string, hotWallets []string, ledgerIndex string) (*types.GatewayBalancesResult, error) {
-	result, err := a.svc.GetGatewayBalances(ctx, account, hotWallets, ledgerIndex)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	var balances map[string][]types.CurrencyBalance
-	if result.Balances != nil {
-		balances = make(map[string][]types.CurrencyBalance)
-		for acct, bals := range result.Balances {
-			rpcBals := make([]types.CurrencyBalance, len(bals))
-			for i, b := range bals {
-				rpcBals[i] = types.CurrencyBalance{
-					Currency: b.Currency,
-					Value:    b.Value,
-				}
-			}
-			balances[acct] = rpcBals
-		}
-	}
-
-	var frozenBalances map[string][]types.CurrencyBalance
-	if result.FrozenBalances != nil {
-		frozenBalances = make(map[string][]types.CurrencyBalance)
-		for acct, bals := range result.FrozenBalances {
-			rpcBals := make([]types.CurrencyBalance, len(bals))
-			for i, b := range bals {
-				rpcBals[i] = types.CurrencyBalance{
-					Currency: b.Currency,
-					Value:    b.Value,
-				}
-			}
-			frozenBalances[acct] = rpcBals
-		}
-	}
-
-	var assets map[string][]types.CurrencyBalance
-	if result.Assets != nil {
-		assets = make(map[string][]types.CurrencyBalance)
-		for acct, bals := range result.Assets {
-			rpcBals := make([]types.CurrencyBalance, len(bals))
-			for i, b := range bals {
-				rpcBals[i] = types.CurrencyBalance{
-					Currency: b.Currency,
-					Value:    b.Value,
-				}
-			}
-			assets[acct] = rpcBals
-		}
-	}
-
-	return &types.GatewayBalancesResult{
-		Account:        result.Account,
-		Obligations:    result.Obligations,
-		Balances:       balances,
-		FrozenBalances: frozenBalances,
-		Assets:         assets,
-		Locked:         result.Locked,
-		LedgerIndex:    result.LedgerIndex,
-		LedgerHash:     result.LedgerHash,
-		Validated:      result.Validated,
-	}, nil
-}
-
-// GetDepositAuthorized checks if a source account is authorized to deposit to a destination account
-func (a *LedgerServiceAdapter) GetDepositAuthorized(ctx context.Context, sourceAccount string, destinationAccount string, ledgerIndex string, credentials []string) (*types.DepositAuthorizedResult, error) {
-	result, err := a.svc.GetDepositAuthorized(ctx, sourceAccount, destinationAccount, ledgerIndex, credentials)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.DepositAuthorizedResult{
-		SourceAccount:      result.SourceAccount,
-		DestinationAccount: result.DestinationAccount,
-		DepositAuthorized:  result.DepositAuthorized,
-		LedgerIndex:        result.LedgerIndex,
-		LedgerHash:         result.LedgerHash,
-		Validated:          result.Validated,
-	}, nil
-}
-
-// GetNFTBuyOffers retrieves buy offers for an NFToken
-func (a *LedgerServiceAdapter) GetNFTBuyOffers(ctx context.Context, nftID [32]byte, ledgerIndex string, limit uint32, marker string) (*types.NFTOffersResult, error) {
-	result, err := a.svc.GetNFTBuyOffers(ctx, nftID, ledgerIndex, limit, marker)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	offers := make([]types.NFTOfferInfo, len(result.Offers))
-	for i, offer := range result.Offers {
-		offers[i] = types.NFTOfferInfo{
-			NFTOfferIndex: offer.NFTOfferIndex,
-			Flags:         offer.Flags,
-			Owner:         offer.Owner,
-			Amount:        offer.Amount,
-			Destination:   offer.Destination,
-			Expiration:    offer.Expiration,
-		}
-	}
-
-	return &types.NFTOffersResult{
-		NFTID:       result.NFTID,
-		Offers:      offers,
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Validated:   result.Validated,
-		Limit:       result.Limit,
-		Marker:      result.Marker,
-	}, nil
-}
-
-// SimulateTransaction runs a transaction against a snapshot without committing
-func (a *LedgerServiceAdapter) SimulateTransaction(txJSON []byte) (*types.SubmitResult, error) {
-	transaction, err := tx.ParseJSON(txJSON)
-	if err != nil {
-		return malformedSubmitResult(), nil
-	}
-
-	result, err := a.svc.SimulateTransaction(transaction)
-	if err != nil {
-		return internalSubmitResult(), nil
-	}
-
-	out := &types.SubmitResult{
-		EngineResult:           result.Result.String(),
-		EngineResultCode:       int(result.Result),
-		EngineResultMessage:    result.Message,
-		Applied:                result.Applied,
-		Fee:                    result.Fee,
-		CurrentLedger:          result.CurrentLedger,
-		CurrentLedgerCloseTime: result.CurrentLedgerCloseTime,
-		ValidatedLedger:        result.ValidatedLedger,
-	}
-	if result.Metadata != nil {
-		blob, serErr := tx.SerializeMetadata(result.Metadata)
-		if serErr != nil {
-			return nil, fmt.Errorf("serialize metadata: %w", serErr)
-		}
-		out.Metadata = &types.SubmitMetadata{
-			JSON: result.Metadata,
-			Blob: blob,
-		}
-	}
-	return out, nil
-}
-
-// TransactionRules returns the rules used by the ledger service for the
-// current open ledger.
-func (a *LedgerServiceAdapter) TransactionRules() *amendment.Rules {
-	return a.svc.TransactionRules()
-}
-
-func (a *LedgerServiceAdapter) GetAutofillFee(txJSON []byte, unlimited bool, mult, div int) (uint64, error) {
-	// rippled getTxFee falls back to reference_fee on any parse failure
-	// (TransactionSign.cpp:790-821). A nil parsedTx triggers the
-	// equivalent baseFee fallback in computeBaseFeeForTx so the later
-	// structural-check passes can surface the real error.
-	parsedTx, _ := tx.ParseJSON(txJSON)
-	return a.svc.GetAutofillFee(parsedTx, unlimited, mult, div)
-}
-
-func (a *LedgerServiceAdapter) GetAutofillSequence(account string, hasTicketSequence bool) (uint32, error) {
-	return a.svc.GetAutofillSequence(account, hasTicketSequence)
-}
-
-// IsAmendmentBlocked returns true if the server is blocked by unsupported amendments
-func (a *LedgerServiceAdapter) IsAmendmentBlocked() bool {
-	return a.svc.IsAmendmentBlocked()
-}
-
-// Table exposes the live amendment table for RPC introspection
-// (feature command, server_info warnings). May be nil.
-func (a *LedgerServiceAdapter) Table() *amendment.Table {
-	return a.svc.Table()
-}
-
-// SetAmendmentVote records an operator veto/upvote and persists it.
-func (a *LedgerServiceAdapter) SetAmendmentVote(ctx context.Context, id [32]byte, vetoed bool) error {
-	return a.svc.SetAmendmentVote(ctx, id, vetoed)
-}
-
-// GetNFTSellOffers retrieves sell offers for an NFToken
-func (a *LedgerServiceAdapter) GetNFTSellOffers(ctx context.Context, nftID [32]byte, ledgerIndex string, limit uint32, marker string) (*types.NFTOffersResult, error) {
-	result, err := a.svc.GetNFTSellOffers(ctx, nftID, ledgerIndex, limit, marker)
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert service result to RPC result
-	offers := make([]types.NFTOfferInfo, len(result.Offers))
-	for i, offer := range result.Offers {
-		offers[i] = types.NFTOfferInfo{
-			NFTOfferIndex: offer.NFTOfferIndex,
-			Flags:         offer.Flags,
-			Owner:         offer.Owner,
-			Amount:        offer.Amount,
-			Destination:   offer.Destination,
-			Expiration:    offer.Expiration,
-		}
-	}
-
-	return &types.NFTOffersResult{
-		NFTID:       result.NFTID,
-		Offers:      offers,
-		LedgerIndex: result.LedgerIndex,
-		LedgerHash:  result.LedgerHash,
-		Validated:   result.Validated,
-		Limit:       result.Limit,
-		Marker:      result.Marker,
-	}, nil
-}
-
-// GetClosedLedgerView returns a read-only view of the last closed ledger
-// for use by pathfinding and other operations that need direct state access.
-func (a *LedgerServiceAdapter) GetClosedLedgerView() (types.LedgerStateView, error) {
-	l := a.svc.GetClosedLedger()
-	if l == nil {
-		return nil, fmt.Errorf("no closed ledger available")
-	}
-	return l, nil
-}
-
-// GetLedgerViewBySeq returns a state view of the ledger with the given
-// sequence, plus its metadata reader.
-func (a *LedgerServiceAdapter) GetLedgerViewBySeq(seq uint32) (types.LedgerStateView, types.LedgerReader, error) {
-	l, err := a.svc.GetLedgerBySequence(seq)
-	if err != nil {
-		return nil, nil, err
-	}
-	return l, &ledgerReaderAdapter{l: l}, nil
-}
-
-// GetLedgerViewByHash returns a state view of the ledger with the given
-// hash, plus its metadata reader.
-func (a *LedgerServiceAdapter) GetLedgerViewByHash(hash [32]byte) (types.LedgerStateView, types.LedgerReader, error) {
-	l, err := a.svc.GetLedgerByHash(hash)
-	if err != nil {
-		return nil, nil, err
-	}
-	return l, &ledgerReaderAdapter{l: l}, nil
-}
